requirements.txt
openai>=1.0.0
requests>=2.31.0
nltk
numpy
scikit-learn
textstat
# youdotcom
pandas
numpy
datasets
sacrebleu
bert-score
transformers
matplotlib
seaborn
tqdm
python-dotenv
scikit-learn





